Remove global ID from search queries
This allows the removal of currency conversion logic. Users will still maintain the ability to search from a specific eBay site but the currency will be the default for that particular location. It did not make sense to allow a US customer to search UK listings since if they would ship to the US then the listings would appear on the US site.